--- a/.gitignore
+++ b/.gitignore
@@ -93,3 +93,6 @@ ENV/
 
 # Rope project settings
 .ropeproject
+
+# PyCharm
+.idea