… pynetdicom/tests/test_service_qr.py