#### Reference issue
Issue #650 

#### What changed
Instead of sleeping before every event in `run_reactor()`, only sleep when there are no queued events to process.

This ensures that queued events are processed as fast as possible. The `_run_loop_delay` setting now only affects the case where there are no queued events (i.e. the system is idle).

After the performance optimization, `pynetdicom` C-STORE SCU is ~23x faster than before when transferring large DICOMs, such as the 2MB `RTImageStorage.dcm` (0.87 sec vs 19.66 sec):

Benchmark BEFORE:
```
1) C-STORE SCU transferred 100 total RTImageStorage.dcm datasets over 1 association in 19.66 s
9) C-STORE SCU/SCP transferred 100 total RTImageStorage.dcm datasets over 1 association in 20.85 s
10) C-STORE DCMTK SCU/SCP transferred 100 total RTImageStorage.dcm datasets over 1 association in 0.20 s
```

Benchmark AFTER:
```
1) C-STORE SCU transferred 100 total RTImageStorage.dcm datasets over 1 association in 0.87 s
9) C-STORE SCU/SCP transferred 100 total RTImageStorage.dcm datasets over 1 association in 1.41 s
10) C-STORE DCMTK SCU/SCP transferred 100 total RTImageStorage.dcm datasets over 1 association in 0.22 s

```

Now `pynetdicom` C-STORE SCU is only ~4x slower than DCMTK (0.87 sec vs. 0.22 sec) instead of ~100x  for `RTImageStorage.dcm`.

The improvement is much more modest for small DICOMs, such as the 39KB `CTImageStorage.dcm` (only ~2x speedup).

The ultrasound images we typically see in our use case are ~200MB and we saw a ~30x speedup on C-STORE SCU with this change.

#### How it was tested
All unit tests pass.  Some unit tests had to be updated to avoid timing-dependent failures.

Also, one unit test fails on `master` even before the changes:
- `pynetdicom/apps/tests/test_common.py:604: AssertionError: assert None == '' where None = getattr((300a, 00c0) Beam Number                         IS: None, 'BeamNumber')`

Seems like there is an extra space in "Beam Number" somewhere?

I also tested the `storescu.py` app with a local instance of OsiriX Lite and it worked properly.

This PR is ready to merge.