@mself 2021-06-21
Note that the unit test `test_scp_cancelled()` also fails on `master` if you set `_run_loop_delay = 0.0001` instead of `_run_loop_delay = 0.001`.  So the failure is not directly caused by the changes in this PR.  This unit tests need to be improved to be less timing-dependent, but I am not sure how to do that.