@scaramallion 2021-06-23
Can you change [these lines](https://github.com/pydicom/pynetdicom/blob/9e8a86d70149e2a23ba2492b705165bf24fc7bdb/pynetdicom/dsutils.py#L213) to:
```python
for elem in iter(ds):
    if elem.VR == 'SQ':
```
To fix the typing error