@scaramallion 2021-06-23
You'll also need to add a test for `DULServiceProvider.stop_dul()` ([these lines](https://app.codecov.io/gh/pydicom/pynetdicom/compare/651/tree/pynetdicom/dul.py#D1L482)) since it's harder to trigger now that the DUL reactor is looping so much faster.