@scaramallion 2021-06-23
The failing test in `test_common` is due to an issue in the current pydicom, it's been fixed in `master`.