@mself 2021-06-24
I don't understand the state machine well enough to create a new unit test for `DULServiceProvider.stop_dul()`.