@scaramallion 2021-06-27
> I don't understand the state machine well enough to create a new unit test for DULServiceProvider.stop_dul().

OK, let me think about it a bit