@mself 2021-06-29
I added a unit test to verify that `stop_dul()` returns `True` when in `Sta1`.  Other unit tests already cover the case where `stop_dul()` returns `False` when in other states.  This should restore the code coverage to 100%.